# docs/swagger.yaml
definitions:
  gin.H:
    additionalProperties: {}
    type: object
...
  github_com_EvolutionAPI_evolution-go_pkg_chatwoot_dto.ChatwootResponse:
    properties:
      accountId:
        type: string
      autoCreate:
        type: boolean
      conversationPending:
        type: boolean
      daysLimitImportMessages:
        type: integer
      enabled:
        type: boolean
      id:
        type: integer
      ignoreJids:
        items:
          type: string
        type: array
      importContacts:
        type: boolean
      importMessages:
        type: boolean
      inboxId:
        type: integer
      instanceId:
        type: string
      logo:
        type: string
      mergeBrazilContacts:
        type: boolean
      nameInbox:
        type: string
      organization:
        type: string
      reopenConversation:
        type: boolean
      signDelimiter:
        type: string
      signMsg:
        type: boolean
      token:
        type: string
      url:
        type: string
    type: object
  github_com_EvolutionAPI_evolution-go_pkg_chatwoot_dto.ChatwootWebhookAttachment:
    properties:
      content_type:
        type: string
      data_url:
        type: string
      file_type:
        type: string
      id:
        type: integer
    type: object
  github_com_EvolutionAPI_evolution-go_pkg_chatwoot_dto.ChatwootWebhookContact:
    properties:
      id:
        type: integer
      identifier:
        type: string
      name:
        type: string
      phone_number:
        type: string
    type: object
  github_com_EvolutionAPI_evolution-go_pkg_chatwoot_dto.ChatwootWebhookConversation:
    properties:
      id:
        type: integer
      meta:
        $ref: '#/definitions/github_com_EvolutionAPI_evolution-go_pkg_chatwoot_dto.ChatwootWebhookConversationMeta'
      status:
        type: string
    type: object
  github_com_EvolutionAPI_evolution-go_pkg_chatwoot_dto.ChatwootWebhookConversationMeta:
    properties:
      channel:
        type: string
      receiver:
        $ref: '#/definitions/github_com_EvolutionAPI_evolution-go_pkg_chatwoot_dto.ChatwootWebhookContact'
      sender:
        $ref: '#/definitions/github_com_EvolutionAPI_evolution-go_pkg_chatwoot_dto.ChatwootWebhookContact'
    type: object
  github_com_EvolutionAPI_evolution-go_pkg_chatwoot_dto.ChatwootWebhookPayload:
    properties:
      attachments:
        items:
          $ref: '#/definitions/github_com_EvolutionAPI_evolution-go_pkg_chatwoot_dto.ChatwootWebhookAttachment'
        type: array
      contact:
        $ref: '#/definitions/github_com_EvolutionAPI_evolution-go_pkg_chatwoot_dto.ChatwootWebhookContact'
      content:
        type: string
      content_type:
        type: string
      conversation:
        $ref: '#/definitions/github_com_EvolutionAPI_evolution-go_pkg_chatwoot_dto.ChatwootWebhookConversation'
      event:
        type: string
      id:
        type: integer
      message_type:
        type: string
      private:
        type: boolean
    type: object
  github_com_EvolutionAPI_evolution-go_pkg_chatwoot_dto.SetChatwootRequest:
    properties:
      accountId:
        type: string
      autoCreate:
        type: boolean
      conversationPending:
        type: boolean
      daysLimitImportMessages:
        type: integer
      enabled:
        type: boolean
      ignoreJids:
        items:
          type: string
        type: array
      importContacts:
        type: boolean
      importMessages:
        type: boolean
      logo:
        type: string
      mergeBrazilContacts:
        type: boolean
      nameInbox:
        type: string
      organization:
        type: string
      reopenConversation:
        type: boolean
      signDelimiter:
        type: string
      signMsg:
        type: boolean
      token:
        type: string
      url:
        type: string
    required:
    - accountId
    - token
    - url
    type: object
...
info:
  contact: {}
  description: Evolution GO - whatsmeow
  title: Evolution GO
  version: "1.0"
paths:
...
  /chatwoot/find/{instance}:
    get:
      description: Returns the persisted Chatwoot settings for the provided instance.
      parameters:
      - description: Instance name
        in: path
        name: instance
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_EvolutionAPI_evolution-go_pkg_chatwoot_dto.ChatwootResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Get Chatwoot configuration for an instance
      tags:
      - Chatwoot
  /chatwoot/set/{instance}:
    post:
      consumes:
      - application/json
      description: Creates or updates the Chatwoot settings associated with the provided
        instance.
      parameters:
      - description: Instance name
        in: path
        name: instance
        required: true
        type: string
      - description: Chatwoot settings
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_EvolutionAPI_evolution-go_pkg_chatwoot_dto.SetChatwootRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_EvolutionAPI_evolution-go_pkg_chatwoot_dto.ChatwootResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Configure Chatwoot integration for an instance
      tags:
      - Chatwoot
  /chatwoot/webhook/{instance}:
    post:
      consumes:
      - application/json
      description: Receives Chatwoot webhook events for a mapped instance and forwards
        supported outgoing messages to WhatsApp.
      parameters:
      - description: Instance name
        in: path
        name: instance
        required: true
        type: string
      - description: Webhook payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_EvolutionAPI_evolution-go_pkg_chatwoot_dto.ChatwootWebhookPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
      summary: Receive webhook from Chatwoot
      tags:
      - Chatwoot
...
swagger: "2.0"